Test and Demo - clean and improve
even if the commit seems big, it's not a major change.
This change is only for the Unit Tests and the Demos.
it's doing some cleaning. More cleaning is still needed in the future.
the main change is that most of Demos are now included in the Unit Test script, which is good for regression test (file: demoTest.cpp)